+++ app/src/main/kotlin/org/fossify/keyboard/helpers/Config.kt
@@ -15,6 +15,10 @@ class Config(context: Context) : BaseConfig(context) {
         get() = prefs.getBoolean(VIBRATE_ON_KEYPRESS, true)
         set(vibrateOnKeypress) = prefs.edit().putBoolean(VIBRATE_ON_KEYPRESS, vibrateOnKeypress).apply()
 
+    var soundOnKeypress: Int
+        get() = prefs.getInt(SOUND_ON_KEYPRESS, SOUND_SYSTEM)
+        set(soundOnKeypress) = prefs.edit().putInt(SOUND_ON_KEYPRESS, soundOnKeypress).apply()
+
     var showPopupOnKeypress: Boolean
         get() = prefs.getBoolean(SHOW_POPUP_ON_KEYPRESS, true)
         set(showPopupOnKeypress) = prefs.edit().putBoolean(SHOW_POPUP_ON_KEYPRESS, showPopupOnKeypress).apply()